backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# AI - using simple OpenAI
openai>=1.10.0

# File processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Security
PyJWT>=2.8.0
bcrypt>=4.1.0

# WebSockets
websockets>=12.0

